Bug 1963853 - Object metric rough c++ support r?janerik
We add a 0-size `Tag` template parameter to ObjectMetric<T> because `object`
metrics with `structure`s that don't have any `type: object` in them can
collide.

In tree this happens with `nimbus_targeting_context::ActiveRolloutsObject`
and `nimbus_targeting_context::ActiveExperimentsObject` which are both
`typedefs` of `nsTArray<nsCString>`.

We're using JSON strings as FFI here because of a lack of pleasant FFI-
compatible optional type, and to minimize copies.

Future optimization would be to not inline the out-of-line definitions of
ObjectMetric<T, Tag>::Set and instead codegen them to a .cpp file. As it stands
this code results in a small-but-more-than-it-needs-to-be increase in binary
size.

I do not consider this to be the final form of support, but it's enough to
unblock my work with the "third-party-modules" ping. "Proper" support would
include a Result<Maybe<T>, nsCString> ObjectMetric<T,Tag>::TestGetValue() impl.

Differential Revision: https://phabricator.services.mozilla.com/D253167