Crashing test
This crashes in the `crash_stack.Set` call, in this assert:

https://searchfox.org/mozilla-central/source/mfbt/JSONWriter.h#335

If I read the code correctly that happens because of a mismatch of the `writer.StartObjectElement` call with the `EndObject` call.
I added the same `crash_stack` object metric to
`toolkit/components/glean/tests/pytest/metrics_test.yaml` and updated
the `metrics_test_output_cpp` to get the code.